requirements.txt
django==1.5
gunicorn
django-honeypot
flufl.password
django-guardian
python_dateutil==2.1
django-evolution
django-ckeditor
django-preferences
django-photologue==2.7
-e git+https://github.com/unomena/unobase.git#egg=unobase
Pillow
python-memcached
South
django-environ==0.4.0